# server/openapi.yaml
openapi: 3.0.3
info:
  title: Sylex Assistant Pro API
  description: |
    Cloud backend API for Sylex Assistant Pro - providing ASR (Automatic Speech Recognition),
    NLU/Chat, and TTS (Text-to-Speech) services for in-vehicle voice assistant.
    
    This API is designed for automotive integration with focus on:
    - Low latency (<300ms TTFB, <1.5s E2E)
    - Safety-first design
    - Multilingual support (DE primary)
    - Streaming support for real-time interaction
  version: 1.0.0
  contact:
    name: Sylex Assistant Pro Team
    email: support@car-assistant-pro.com
  license:
    name: Proprietary
servers:
  - url: https://api.car-assistant-pro.com/v1
    description: Production server
  - url: https://staging-api.car-assistant-pro.com/v1
    description: Staging server
  - url: http://localhost:3000/v1
    description: Local development server
...
paths:
...
  /chat:
    post:
      tags:
        - Chat
      summary: Process natural language query
      description: |
        Processes user text input through NLU pipeline and dialogue manager.
        Returns intent classification, slot values, and assistant response.
      operationId: processChat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
                  description: User input text
                  maxLength: 500
                  example: "Navigate to nearest charging station"
                sessionId:
                  type: string
                  description: Session identifier for context continuity
                  format: uuid
                  example: "550e8400-e29b-41d4-a716-446655440000"
                context:
                  type: object
                  description: Additional context (location, vehicle state, etc.)
                  additionalProperties: true
                  example:
                    location:
                      lat: 52.520008
                      lon: 13.404954
                    vehicleState:
                      batteryLevel: 45
                      speed: 0
                language:
                  type: string
                  default: de-DE
                  example: de-DE
              required:
                - text
      responses:
        '200':
          description: Chat processing successful
          headers:
            X-Request-ID:
              schema:
                type: string
            X-Processing-Time-Ms:
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          headers:
            X-RateLimit-Limit:
              schema:
                type: integer
            X-RateLimit-Remaining:
              schema:
                type: integer
            X-RateLimit-Reset:
              schema:
                type: integer
            Retry-After:
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ChatResponse:
      type: object
      properties:
        requestId:
          type: string
          format: uuid
        sessionId:
          type: string
          format: uuid
        intent:
          type: object
          properties:
            name:
              type: string
              description: Intent identifier
              example: "navigate_to_poi"
            confidence:
              type: number
              format: float
              minimum: 0.0
              maximum: 1.0
              example: 0.92
        slots:
          type: object
          description: Extracted slot values
          additionalProperties: true
          example:
            poi_type: "charging_station"
            modifier: "nearest"
        response:
          type: object
          properties:
            text:
              type: string
              description: Assistant response text
              example: "Navigating to the nearest charging station, 2.3 km away"
            action:
              type: string
              description: Action to execute
              example: "navigate"
            parameters:
              type: object
              description: Action parameters
              additionalProperties: true
        processingTimeMs:
          type: integer
      required:
        - requestId
        - sessionId
        - intent
        - response
        - processingTimeMs
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Error code
              enum:
                - INVALID_REQUEST
                - INVALID_AUDIO_FORMAT
                - AUDIO_TOO_LARGE
                - TEXT_TOO_LONG
                - RATE_LIMIT_EXCEEDED
                - UNAUTHORIZED
                - PROVIDER_ERROR
                - INTERNAL_ERROR
                - SERVICE_UNAVAILABLE
              example: "INVALID_AUDIO_FORMAT"
            message:
              type: string
              description: Human-readable error message
              example: "Unsupported audio format. Supported formats: WAV, MP3, FLAC, OGG"
            details:
              type: object
              description: Additional error details
              additionalProperties: true
            requestId:
              type: string
              format: uuid
              description: Request ID for tracking
          required:
            - code
            - message
      required:
        - error
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
security:
  - BearerAuth: []
...